+++ Alter Eco/Alter Eco/Frontend/ProgressBarView.swift
@@ -20,14 +20,14 @@ struct ProgressBarView: View {
             //text box for league information
             ZStack() {
                RoundedRectangle(cornerRadius: 25, style: .continuous)
-                   .frame(width: CGFloat(screenMeasurements.broadcastedWidth) * 0.9, height: CGFloat(screenMeasurements.broadcastedHeight) / 7)
+                   .frame(width: screenMeasurements.broadcastedWidth * 0.9, height: screenMeasurements.broadcastedHeight / 7)
                    .foregroundColor(Color("fill_colour"))
             
                 VStack() {
                     
                         if ((try! DBMS.retrieveLatestScore()).league != "🌳") {
                             Text("Grow your forest! You have planted \((try! DBMS.retrieveLatestScore()).counter) 🌳 so far, keep earning points to grow a new 🌳!")
-                            .frame(width: CGFloat(screenMeasurements.broadcastedWidth)*0.7, height: CGFloat(screenMeasurements.broadcastedHeight) / 8)
+                            .frame(width: screenMeasurements.broadcastedWidth*0.7, height: screenMeasurements.broadcastedHeight / 8)
                         }
                          
                        // depending on which league user is in, display next one
@@ -35,7 +35,7 @@ struct ProgressBarView: View {
                            Text("Your forest is thriving! You just planted another 🌳, for a total of \((try! DBMS.retrieveLatestScore()).counter)! Congratulations! You're now growing a new sapling.")
                                 .font(.headline)
                                 .fontWeight(.regular)
-                                .frame(width: CGFloat(screenMeasurements.broadcastedWidth)*0.7, height: CGFloat(screenMeasurements.broadcastedHeight)/8)
+                                .frame(width: screenMeasurements.broadcastedWidth*0.7, height: screenMeasurements.broadcastedHeight/8)
                        }
                 }
            }
@@ -47,13 +47,13 @@ struct ProgressBarView: View {
                 
                 ZStack(alignment: .leading) {
                     Rectangle()
-                        .frame(width: CGFloat(screenMeasurements.broadcastedWidth)*0.7, height: CGFloat(screenMeasurements.broadcastedHeight)/45)
+                        .frame(width: screenMeasurements.broadcastedWidth*0.7, height: screenMeasurements.broadcastedHeight/45)
                         .opacity(0.3)
                         .foregroundColor(Color("fill_colour"))
               
                     
                         Rectangle()
-                           .frame(width: (CGFloat(screenMeasurements.broadcastedWidth)*CGFloat(0.7*proportion)), height: CGFloat(screenMeasurements.broadcastedHeight)/45)
+                           .frame(width: screenMeasurements.broadcastedWidth*CGFloat(0.7*proportion), height: screenMeasurements.broadcastedHeight/45)
                             .foregroundColor(Color("graphBars"))
                             .animation(.linear)
                             
@@ -64,7 +64,7 @@ struct ProgressBarView: View {
             }
           Text("\((try! DBMS.retrieveLatestScore()).totalPoints, specifier: "%.0f") / \(POINTS_REQUIRED_FOR_NEXT_LEAGUE, specifier: "%.0f")")
             .font(.body)
-            .padding(.leading, CGFloat(screenMeasurements.broadcastedWidth)/3.5)
+            .padding(.leading, screenMeasurements.broadcastedWidth/3.5)
             
         }
     }
