+++ Alter Eco/Alter Eco/Frontend/ScreenMeasurements.swift
@@ -2,6 +2,6 @@ import Foundation
 import SwiftUI
 
 class ScreenMeasurements: ObservableObject {
-    @Published var broadcastedHeight: Float = Float(UIScreen.main.bounds.height)
-    @Published var broadcastedWidth: Float = Float(UIScreen.main.bounds.width)
+    @Published var broadcastedHeight: CGFloat = CGFloat(UIScreen.main.bounds.height)
+    @Published var broadcastedWidth: CGFloat = CGFloat(UIScreen.main.bounds.width)
 }
